requirements.txt
Flask==2.3.2
peewee==3.16.2
atlassian-python-api
python-dotenv
lxml==4.9.3
beautifulsoup4
pandas==2.2.2
pyuwsgi==2.0.23.post0
uWSGI==2.0.24
pytest
